+++ packages/core/src/ai-sdk/language-model-middleware/v3/index.ts
@@ -1,1 +1,1 @@
-export * from './language-model-v3-billing-middleware.js';
+export * from './language-model-v3-base-billing-middleware.js';
